requirements.txt
alabaster==1.0.0
asgiref==3.7.2
babel==2.17.0
boto3==1.34.69
botocore==1.34.69
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.0.2
django-ckeditor==6.7.2
django-cors-headers==4.3.1
django-js-asset==3.1.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.3.0
gunicorn==21.2.0
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==24.0
pillow==10.2.0
pluggy==1.6.0
psycopg2==2.9.9
Pygments==2.19.1
PyJWT==2.8.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.4
s3transfer==0.10.1
six==1.16.0
snowballstemmer==3.0.1
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.4.4
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2023.4
urllib3==2.2.1
whitenoise==6.6.0
